adding replace
